Factor display.h|cpp out of panels.h|cpp
