+++ src/character_body.cpp
@@ -1,5 +1,6 @@
 #include "avatar.h"
 #include "character.h"
+#include "display.h"
 #include "flag.h"
 #include "game.h"
 #include "map.h"
@@ -8,7 +9,6 @@
 #include "options.h"
 #include "output.h"
 #include "overmapbuffer.h"
-#include "panels.h"
 #include "type_id.h"
 #include "vitamin.h"
 #include "veh_type.h"
