+++ src/faction.cpp
@@ -18,6 +18,7 @@
 #include "coordinates.h"
 #include "cursesdef.h"
 #include "debug.h"
+#include "display.h"
 #include "faction_camp.h"
 #include "game.h"
 #include "game_constants.h"
@@ -31,7 +32,6 @@
 #include "optional.h"
 #include "output.h"
 #include "overmapbuffer.h"
-#include "panels.h"
 #include "pimpl.h"
 #include "point.h"
 #include "skill.h"
