+++ src/panels.h
@@ -36,110 +36,6 @@ namespace catacurses
 class window;
 } // namespace catacurses
 
-// The display namespace contains UI string output and colorization functions
-// Some return plain strings or translations, some return a (string, color) pair,
-// and some return a string with colorization tags embedded.
-namespace display
-{
-// Functions returning plain strings
-// Current moon phase, ex. "Full moon", "Waxing crescent"
-std::string get_moon();
-// Current moon phase as ascii-art, ex. "(   )", "(  ))"
-std::string get_moon_graphic();
-// Current date, in terms of day within season, ex. "Summer, day 17"
-std::string date_string();
-// Current approximate time of day, ex. "Early morning", "Around dusk"
-std::string time_approx();
-// Exact time if character has a watch, approx time if aboveground, "???" if unknown/underground
-std::string time_string( const Character &u );
-
-// Temperature at character location, if they have a thermometer
-std::string get_temp( const Character &u );
-// Change in character body temperature, ex. "(Rising)", "(Falling!!)"
-std::string temp_delta_string( const Character &u );
-
-// Text descriptor for given activity level, ex. "Light", "Brisk", "Extreme"
-std::string activity_level_str( float level );
-// gets the malus string for character's current activity level, like "+ 25%"
-std::string activity_malus_str( const Character &u );
-// gets the description, printed in player_display, related to your current bmi
-std::string weight_long_description( const Character &u );
-
-// For vehicle being driven or remotely piloted by character
-// Azimuth (heading) in degrees
-std::string vehicle_azimuth_text( const Character &u );
-// Vehicle target/current cruise velocity (and units) with engine strain color
-std::pair<std::string, nc_color> vehicle_cruise_text_color( const Character &u );
-// Vehicle percent of fuel remaining for currently running engine
-std::pair<std::string, nc_color> vehicle_fuel_percent_text_color( const Character &u );
-
-// Functions returning (text, color) pairs
-std::pair<translation, nc_color> weariness_text_color( size_t weariness );
-std::pair<std::string, nc_color> weariness_text_color( const Character &u );
-std::pair<std::string, nc_color> weary_malus_text_color( const Character &u );
-std::pair<std::string, nc_color> activity_text_color( const Character &u );
-std::pair<std::string, nc_color> thirst_text_color( const Character &u );
-std::pair<std::string, nc_color> hunger_text_color( const Character &u );
-std::pair<std::string, nc_color> weight_text_color( const Character &u );
-std::pair<std::string, nc_color> fatigue_text_color( const Character &u );
-std::pair<std::string, nc_color> health_text_color( const Character &u );
-std::pair<std::string, nc_color> pain_text_color( const Creature &c );
-std::pair<std::string, nc_color> pain_text_color( const Character &u );
-// Change in character body temperature, as colorized arrows
-std::pair<std::string, nc_color> temp_delta_arrows( const Character &u );
-// Character morale, as a color-coded ascii emoticon face
-std::pair<std::string, nc_color> morale_face_color( const avatar &u );
-// Helpers for morale_face_color
-std::pair<std::string, nc_color> morale_emotion( const int morale_cur, const mood_face &face );
-
-// Current movement mode and color, as single letter or full word
-std::pair<std::string, nc_color> move_mode_letter_color( const Character &u );
-std::pair<std::string, nc_color> move_mode_text_color( const Character &u );
-// Current body part status (bleeding, bitten, infected) phrase, fully colorized
-std::string colorized_bodypart_status_text( const Character &u, const bodypart_id &bp );
-
-std::pair<std::string, nc_color> temp_text_color( const Character &u );
-std::pair<std::string, nc_color> power_text_color( const Character &u );
-std::pair<std::string, nc_color> mana_text_color( const Character &you );
-std::pair<std::string, nc_color> str_text_color( const Character &p );
-std::pair<std::string, nc_color> dex_text_color( const Character &p );
-std::pair<std::string, nc_color> int_text_color( const Character &p );
-std::pair<std::string, nc_color> per_text_color( const Character &p );
-std::pair<std::string, nc_color> safe_mode_text_color( const bool classic_mode );
-std::pair<std::string, nc_color> wind_text_color( const Character &u );
-std::pair<std::string, nc_color> weather_text_color( const Character &u );
-
-// Get visible threats by cardinal direction - Already colorized
-std::string colorized_compass_text( const cardinal_direction dir, int width );
-std::string colorized_compass_legend_text( int width, int height );
-
-// Define color for displaying the body temperature
-nc_color bodytemp_color( const Character &u, const bodypart_id &bp );
-// Returns color which this limb would have in healing menus
-nc_color limb_color( const Character &u, const bodypart_id &bp, bool bleed, bool bite,
-                     bool infect );
-// Return strings for all statuses affecting body part (bleeding, bitten, bandaged, etc.)
-std::vector<std::pair<std::string, nc_color>> bodypart_status_colors( const Character &u,
-        const bodypart_id &bp );
-
-// Color for displaying the given encumbrance level
-nc_color encumb_color( const int level );
-
-// Color name for given irradiation level, used for radiation badge description
-std::string rad_badge_color_name( const int rad );
-// Color for displaying the given irradiation level
-nc_color rad_badge_color( const int rad );
-// Highlighted badge color for character's radiation badge, if they have one
-std::pair<std::string, nc_color> rad_badge_text_color( const Character &u );
-
-// Functions returning colorized string
-// gets the string that describes your weight
-std::string weight_string( const Character &u );
-
-// Prints a list of nearby monsters
-void print_mon_info( const avatar &u, const catacurses::window &, int hor_padding = 0,
-                     bool compact = false );
-} // namespace display
 
 namespace overmap_ui
 {
