+++ src/player_display.cpp
@@ -18,6 +18,7 @@
 #include "color.h"
 #include "cursesdef.h"
 #include "debug.h"
+#include "display.h"
 #include "effect.h"
 #include "enum_conversions.h"
 #include "game.h"
@@ -26,7 +27,6 @@
 #include "mutation.h"
 #include "options.h"
 #include "output.h"
-#include "panels.h"
 #include "pimpl.h"
 #include "pldata.h"
 #include "profession.h"
