+++ src/suffer.cpp
@@ -21,6 +21,7 @@
 #include "character.h"
 #include "creature.h"
 #include "debug.h"
+#include "display.h"
 #include "effect.h"
 #include "enums.h"
 #include "event.h"
@@ -46,7 +47,6 @@
 #include "options.h"
 #include "output.h"
 #include "overmapbuffer.h"
-#include "panels.h"
 #include "pimpl.h"
 #include "player_activity.h"
 #include "pldata.h"
