+++ tests/char_biometrics_test.cpp
@@ -8,11 +8,11 @@
 #include "calendar.h"
 #include "cata_catch.h"
 #include "creature.h"
+#include "display.h"
 #include "game_constants.h"
 #include "monster.h"
 #include "options.h"
 #include "output.h"
-#include "panels.h"
 #include "player_helpers.h"
 #include "string_formatter.h"
 #include "type_id.h"
